Moves all top-level cmd utils into a utils package.
